requirements.txt
yfinance
alpaca-trade-api>=3.0.0
requests
python-telegram-bot
python-dotenv
